test_app/requirements.txt
openai==1.93.1
python-dotenv==1.0.0
fastapi==0.108.0
uvicorn[standard]==0.25.0
tiktoken==0.5.2

sentence-transformers>=2.6.1
huggingface-hub>=0.30.0,<1.0
qdrant-client==1.9.0
pandas

